changed to float

+++ ee250/final/rpi_pub_sub.py
@@ -81,7 +81,7 @@ def weather_sensor_callback(client, userdata, message):
     print("Weather from the sensor: " + str(message.payload, 'utf-8') + "F")
     temp = str(message.payload, 'utf-8')
     setText_norefresh(temp) # output to the lcd screen
-    difference = int(server_weather) - int(temp)
+    difference = float(server_weather) - float(temp)
     if difference < 0:
         difference = difference*-1
